s/my-quickfiles/quickfiles/ and added missing tests
