--- tests/unit/my-quickfiles/route-test.js
@@ -1,11 +1,0 @@
-import { moduleFor, test } from 'ember-qunit';
-
-moduleFor('route:my-quickfiles', 'Unit | Route | my quickfiles', {
-  // Specify the other units that are required for this test.
-  // needs: ['controller:foo']
-});
-
-test('it exists', function(assert) {
-  let route = this.subject();
-  assert.ok(route);
-});
