+++ tests/unit/user-quickfiles/controller-test.js
@@ -1,0 +1,12 @@
+import { moduleFor, test } from 'ember-qunit';
+
+moduleFor('controller:user-quickfiles', 'Unit | Controller | user quickfiles', {
+    // Specify the other units that are required for this test.
+    // needs: ['controller:foo']
+});
+
+// Replace this with your real tests.
+test('it exists', function(assert) {
+    let controller = this.subject();
+    assert.ok(controller);
+});
