+++ tests/unit/user-quickfiles/route-test.js
@@ -1,0 +1,11 @@
+import { moduleFor, test } from 'ember-qunit';
+
+moduleFor('route:user-quickfiles', 'Unit | Route | user quickfiles', {
+    // Specify the other units that are required for this test.
+    // needs: ['controller:foo']
+});
+
+test('it exists', function(assert) {
+    let route = this.subject();
+    assert.ok(route);
+});
